flight-booking/booking_service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.3
pydantic==2.10.6
email-validator==2.2.0
